+++ src/main/kotlin/com/mairwunnx/projectessentials/permissions/permissions/PermissionBase.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.UnstableDefault
 import org.apache.logging.log4j.LogManager
 import java.io.File
 
-@UseExperimental(UnstableDefault::class)
+@OptIn(UnstableDefault::class)
 internal object PermissionBase {
     private val logger = LogManager.getLogger()
     internal var permissionData = PermissionModel()
@@ -18,7 +18,8 @@ internal object PermissionBase {
             logger.warn("Permission config not exist! creating it now!")
             File(MOD_CONFIG_FOLDER).mkdirs()
             val defaultConfig = jsonInstance.stringify(
-                PermissionModel.serializer(), permissionData
+                PermissionModel.serializer(),
+                permissionData
             )
             File(permissionConfig).writeText(defaultConfig)
         }
@@ -32,7 +33,8 @@ internal object PermissionBase {
         val permissionConfig = MOD_CONFIG_FOLDER + File.separator + "permissions.json"
         File(MOD_CONFIG_FOLDER).mkdirs()
         val permConfig = jsonInstance.stringify(
-            PermissionModel.serializer(), permissionData
+            PermissionModel.serializer(),
+            permissionData
         )
         File(permissionConfig).writeText(permConfig)
     }